Adding Common folder and some smaller refactorings
