+++ common/pom.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>io.streamzi.cloudevent-flow</groupId>
+        <artifactId>parent</artifactId>
+        <version>0.0.4-SNAPSHOT</version>
+    </parent>
+    <artifactId>common</artifactId>
+    <name>cloudevent-flow-${project.artifactId}</name>
+    <packaging>jar</packaging>
+
+</project>
