+++ container/pom.xml
@@ -18,6 +18,12 @@
         </dependency>
 
         <dependency>
+            <groupId>io.streamzi.cloudevent-flow</groupId>
+            <artifactId>common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
             <groupId>eu.infomas</groupId>
             <artifactId>annotation-detector</artifactId>
             <version>3.0.5</version>
