+++ container/src/main/java/io/streamzi/openshift/dataflow/container/CloudEventInput.java
@@ -2,8 +2,9 @@ package io.streamzi.openshift.dataflow.container;
 
 import io.streamzi.openshift.dataflow.annotations.CloudEventConsumer;
 import io.streamzi.openshift.dataflow.annotations.ObjectType;
-import io.streamzi.openshift.dataflow.container.config.EnvironmentResolver;
 import io.streamzi.openshift.dataflow.model.ProcessorConstants;
+import io.streamzi.openshift.dataflow.utils.EnvironmentResolver;
+
 import java.lang.reflect.Method;
 import java.util.logging.Logger;
