+++ watcher/src/main/java/io/streamzi/openshift/FlowController.java
@@ -113,10 +113,9 @@ public class FlowController {
 
                     });
 
-            System.out.println("Done");
+            logger.info("Done");
 
         } catch (Exception e) {
-            e.printStackTrace();
             logger.log(Level.SEVERE, "Error parsing JSON flow data: " + e.getMessage(), e);
         }
     }
